integration with floor and tables
