+++ src/floor/Table.h
@@ -74,7 +74,6 @@ public:
      */
     virtual int capacity() const = 0;
     virtual uint32_t id() const = 0;
-
     virtual std::string toString() const = 0;
 
 private:
